refactor(types): split into separate files
