+++ crates/type/src/directory.rs
@@ -1,0 +1,33 @@
+// SPDX-FileCopyrightText: 2022 Profian Inc. <opensource@profian.com>
+// SPDX-License-Identifier: Apache-2.0
+
+use super::entry::Entry;
+
+use std::collections::BTreeMap;
+use std::ops::{Deref, DerefMut};
+
+use serde::{Deserialize, Serialize};
+
+/// A directory
+///
+/// A directory is simply a sorted name to `Entry` map.
+#[derive(Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
+pub struct Directory(BTreeMap<String, Entry>);
+
+impl Directory {
+    pub const TYPE: &'static str = "application/vnd.drawbridge.directory.v1+json";
+}
+
+impl Deref for Directory {
+    type Target = BTreeMap<String, Entry>;
+
+    fn deref(&self) -> &Self::Target {
+        &self.0
+    }
+}
+
+impl DerefMut for Directory {
+    fn deref_mut(&mut self) -> &mut Self::Target {
+        &mut self.0
+    }
+}
